refactor(brand): replace BrandProvider context with plain module + Locale/Direction enums
- Extract branding config to src/config/branding.ts (BrandConfig, AuthPageConfig,
  I18nConfig interfaces + parsers + pre-parsed singletons + useBrand() + initBranding())
- Add Locale and Direction enums replacing string literals
- Delete src/context/BrandContext.tsx — no provider wrapper needed for static config
- Move DOM side effects (CSS var + favicon) to initBranding(), called in main.tsx
- Update all useBrand() imports to @/config/branding
- Update initI18n signature to accept Direction enum
