+++ src/main.tsx
@@ -6,12 +6,14 @@ import SentryProvider from './core/services/sentry/SentryProvider.tsx';
 import VercelSpeedInsights from './core/services/vercel/vercel.tsx';
 import { config } from './config/config.ts';
 import { registerWebMCPTools } from './agent/webmcp.ts';
-import { BrandProvider } from './context/BrandContext.tsx';
+import { initBranding } from './config/branding.ts';
 import { initI18n } from './i18n/index.ts';
 
 registerWebMCPTools();
 
 (async () => {
+	initBranding();
+
 	try {
 		await initI18n(config.i18n.locale, config.i18n.direction);
 	} catch (err) {
@@ -19,19 +21,17 @@ registerWebMCPTools();
 	}
 
 	ReactDOM.createRoot(document.getElementById('root')!).render(
-		<BrandProvider>
-			<div>
-				{config.app.isProd ? (
-					<SentryProvider>
-						<PosthogProvider>
-							<App />
-							<VercelSpeedInsights />
-						</PosthogProvider>
-					</SentryProvider>
-				) : (
-					<App />
-				)}
-			</div>
-		</BrandProvider>,
+		<div>
+			{config.app.isProd ? (
+				<SentryProvider>
+					<PosthogProvider>
+						<App />
+						<VercelSpeedInsights />
+					</PosthogProvider>
+				</SentryProvider>
+			) : (
+				<App />
+			)}
+		</div>,
 	);
 })();
